backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-base64==2.0
Pillow==9.4.0
PyJWT==2.6.0
requests==2.28.2
sqlparse==0.4.3
django-filter==2.4.0
webcolors==1.11.1
django-cors-headers==3.14.0
django-colorfield
